Move task to a dedicated project
