+++ server/pom.xml
@@ -98,6 +98,7 @@
         <module>queue/queue-file</module>
         <module>queue/queue-jms</module>
 
+        <module>task</module>
         <module>testing</module>
     </modules>
