+++ server/task/src/main/java/org/apache/james/task/Task.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.cassandra.mail.task;
+package org.apache.james.task;
 
 import java.util.Arrays;
 import java.util.Objects;
